Export custom plugin to seperate file
